rename sqlutil package
