+++ vectorstore/psql/psql.go
@@ -23,7 +23,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/henomis/phero/internal/sqlutil"
+	sqlutil "github.com/henomis/phero/internal/sql"
 	"github.com/henomis/phero/vectorstore"
 )
